+++ src/data/stores/ExecutionWindowStore.ts
@@ -1,0 +1,8 @@
+import { YAMLStore } from './BaseStore';
+import { ExecutionWindow } from '../types';
+
+export class ExecutionWindowStore extends YAMLStore<ExecutionWindow> {
+  constructor(workspaceRoot?: string) {
+    super('workforce', 'executionWindows.yml', 'executionWindows', workspaceRoot);
+  }
+}